aircox_cms/requirements.txt
Django>=1.9.0
django-taggit>=0.12.1
easy_thumbnails

